+++ Ting/Models/Post.swift
@@ -32,55 +32,3 @@ struct Post: Identifiable, Codable {
     var currentStatus: String? // 현재상태 - "취준", "현업", "경력", "기타"
 }
 
-/// 신고 모델 구조체
-struct Report: Codable {
-    @DocumentID var id: String?
-    
-    let reportReason: String      // 신고 사유
-    let reportDetails: String     // 상세 내용
-    let title: String            // 게시글 제목
-    let reporterNickname: String  // 신고자 닉네임
-    let creationTime: String      // 작성 시간
-    let nickname: String          // 신고받는 사람 닉네임
-    
-    var toDictionary: [String: Any] {
-        return [
-            "reportReason": reportReason,
-            "reportDetails": reportDetails,
-            "title": title,
-            "reporterNickname": reporterNickname,
-            "creationTime": creationTime,
-            "nickname": nickname
-        ]
-    }
-}
-
-
-class ReportManager {
-    static let shared = ReportManager()
-    private let db = Firestore.firestore()
-    private let reportCollection = "reports"
-    
-    private init() {}
-    
-    func uploadReport(_ report: Report, completion: @escaping (Result<Void, Error>) -> Void) {
-        do {
-            let data = try Firestore.Encoder().encode(report)
-            db.collection(reportCollection).addDocument(data: data) { error in
-                if let error = error {
-                    completion(.failure(error))
-                } else {
-                    completion(.success(()))
-                }
-            }
-        } catch {
-            completion(.failure(error))
-        }
-    }
-    
-    func getCurrentTime() -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy/MM/dd"
-        return formatter.string(from: Date())
-    }
-}
